backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
python-multipart==0.0.9
pydantic[email]
